requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
Pillow>=10.0.0
